+++ tests/testthat/test-score_forecasts.R
@@ -154,6 +154,19 @@ test_that("(wide version) calculated scores that are in scores vector should hav
   
 })
 
+
+test_that("WIS calcultaion with point forecast use_median_as_point FALSE",{
+  
+})
+
+test_that("WIS calcultaion with point forecast use_median_as_point TRUE",{
+  
+})
+
+test_that("WIS calcultaion without point forecast use_median_as_point FALSE",{
+  # should error
+})
+
 test_that("abs error is correct, point forecast only", {
   y <- c(1, -15, 22)
   
@@ -196,15 +209,14 @@ test_that("abs error is correct, point forecast only", {
     stringsAsFactors = FALSE
   )
   
-  expect_error(actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth))
+  actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth, use_median_as_point = FALSE)
   
-  # expected <- abs(y - point_forecast)
-  # 
-  # expect_equal(actual$abs_error, expected)
-  
+  expected <- abs(y - point_forecast)
+
+  expect_equal(actual$abs_error, expected)
 })
 
-test_that("abs error is correct, point and median forecasts different", {
+test_that("abs error is correct, point and median forecasts different, use_median_as_point is FALSE", {
   y <- c(1, -15, 22)
   forecast_quantiles_matrix <- rbind(
     c(-1, 0, 1, 2, 3),
@@ -255,7 +267,65 @@ test_that("abs error is correct, point and median forecasts different", {
     stringsAsFactors = FALSE
   )
   
-  actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth)
+  actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth, use_median_as_point = FALSE)
+  
+  expected <- abs(y - point_forecast)
+  
+  expect_equal(actual$abs_error, expected)
+})
+
+test_that("abs error is correct, point and median forecasts different, use_median_as_point is TRUE", {
+  y <- c(1, -15, 22)
+  forecast_quantiles_matrix <- rbind(
+    c(-1, 0, 1, 2, 3),
+    c(-2, 1, 2, 2, 4),
+    c(-2, 0, 3, 3, 4))
+  forecast_quantile_probs <- c(0.1, 0.25, 0.5, 0.75, 0.9)
+  forecast_quantiles_matrix <- forecast_quantiles_matrix[, 3, drop = FALSE]
+  forecast_quantile_probs <- forecast_quantile_probs[3]
+  
+  target_end_dates <- lubridate::ymd(20200101) + c(7, 14, 7)
+  horizons <- c("1", "2", "1")
+  locations <- c("01", "01", "02")
+  target_variables <- rep("inc death", length(y))
+  
+  forecast_target_end_dates <-
+    rep(target_end_dates, times = 1 + ncol(forecast_quantiles_matrix))
+  forecast_horizons <- rep(horizons, times = 1 + ncol(forecast_quantiles_matrix))
+  forecast_locations <- rep(locations, times = 1 + ncol(forecast_quantiles_matrix))
+  forecast_target_variables <-
+    rep(target_variables, times = 1 + ncol(forecast_quantiles_matrix))
+  forecast_quantile_probs <- rep(forecast_quantile_probs, each = length(y))
+  forecast_quantiles <- forecast_quantiles_matrix
+  dim(forecast_quantiles) <- prod(dim(forecast_quantiles))
+  
+  point_forecast <- c(5,6,7)
+  
+  test_truth <- data.frame(
+    model = rep("truth_source", length(y)),
+    target_variable = target_variables,
+    target_end_date = target_end_dates,
+    location = locations,
+    value = y,
+    stringsAsFactors = FALSE
+  )
+  
+  n_forecasts <- length(point_forecast) + length(forecast_quantiles)
+  test_forecasts <- data.frame(
+    model = rep("m1", n_forecasts),
+    forecast_date = rep(lubridate::ymd("20200101"), n_forecasts),
+    location = forecast_locations,
+    horizon = forecast_horizons,
+    temporal_resolution = rep("wk", n_forecasts),
+    target_variable = forecast_target_variables,
+    target_end_date = forecast_target_end_dates,
+    type = c(rep("point",length(point_forecast)),rep("quantile", length(forecast_quantiles))),
+    quantile = c(rep(NA,length(point_forecast)),forecast_quantile_probs),
+    value = c(point_forecast,forecast_quantiles),
+    stringsAsFactors = FALSE
+  )
+  
+  actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth, use_median_as_point = TRUE)
   
   expected <- abs(y - c(1,2,3))
   
@@ -313,11 +383,12 @@ test_that("abs error is correct, point and median forecasts same", {
     stringsAsFactors = FALSE
   )
   
-  actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth)
+  actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth, use_median_as_point = FALSE)
   
   expected <- abs(y - point_forecast)
   
   expect_equal(actual$abs_error, expected)
+
 })
 
 test_that("wis is correct, median only", {
@@ -369,7 +440,8 @@ test_that("wis is correct, median only", {
     stringsAsFactors = FALSE
   )
 
-  actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth)
+  actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth, 
+                            use_median_as_point = TRUE)
 
   expected <- abs(y - forecast_quantiles_matrix[, 1])
 
@@ -426,7 +498,8 @@ test_that("wis is correct, 1 interval only", {
     stringsAsFactors = FALSE
   )
 
-  actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth)
+  actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth, 
+                            use_median_as_point = TRUE)
 
   alpha1 <- 0.2
   expected <- (forecast_quantiles_matrix[, 2] - forecast_quantiles_matrix[, 1]) * (alpha1 / 2) +
@@ -483,8 +556,8 @@ test_that("wis is correct, 2 intervals and median", {
     stringsAsFactors = FALSE
   )
 
-  actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth)
-
+  actual <- score_forecasts(forecasts = test_forecasts, truth = test_truth, 
+                            use_median_as_point = TRUE)
   alpha1 <- 0.2
   alpha2 <- 0.5
   expected <- (1 / 2.5) * (